fix: build release-tool on Windows and wire BDD matrix entry (#923)
Two CI failures from the PR-3 push, both root-cause "I forgot another
guard while adding the new BDD tag and the new syscall":

1. Cross-Platform (. on windows-latest) failed because the workspace
   pulls cmd/release-tool into the same build graph as core, and
   readAllowedFile referenced syscall.O_NOFOLLOW + syscall.ELOOP
   which are POSIX-only. Splits the helper:
   - read_allowed_unix.go (//go:build !windows) keeps the
     O_NOFOLLOW + IsRegular() pair as the security gate
   - read_allowed_windows.go (//go:build windows) returns an
     explicit "not implemented on Windows" error so an operator
     who somehow runs the binary there gets an immediate refusal
     rather than a silently follows-symlinks fallback
   - read_allowed_windows_test.go locks the refusal contract

2. BDD Coverage Verification scenario-count guard caught the
   release-tool feature not being executed by any matrix runner:
   adds the matching matrix entry to .github/workflows/ci.yml so
   the 8 @release-tool scenarios run alongside core/file/syslog/etc.

Refs #918
